Rename lcm.aggregators to lcm.temporal_aggregation
