+++ src/lcm/__init__.py
@@ -53,7 +53,6 @@ from _lcm.variables import (  # noqa: E402
 )
 from _lcm.version import __version__  # noqa: E402
 from lcm.ages import AgeGrid  # noqa: E402
-from lcm.aggregators import H_epstein_zin, H_linear  # noqa: E402
 from lcm.certainty_equivalent import (  # noqa: E402
     CertaintyEquivalent,
     PowerMean,
@@ -97,6 +96,7 @@ from lcm.regime import (  # noqa: E402
 from lcm.result import SimulationResult  # noqa: E402
 from lcm.solvers import DCEGM, GridSearch  # noqa: E402
 from lcm.taste_shocks import ExtremeValueTasteShocks  # noqa: E402
+from lcm.temporal_aggregation import H_epstein_zin, H_linear  # noqa: E402
 from lcm.transition import fixed_transition  # noqa: E402
 
 # Modules with TYPE_CHECKING-only forward references expose a
